feat: MongoDB support
- Added support for MongoDB as a backing storage provider.
- Refactored some code which will allow support for other storage implementations to be added more easily in the future.
- As a result of this refactoring, the in-memory storage provider can now be used for vault encrypted indexes + querying.
- Added a TODO marker to the OpenAPI demo script for an existing issue.
