+++ cmd/edv-rest/go.mod
@@ -11,15 +11,17 @@ require (
 	github.com/google/tink/go v1.6.1-0.20210519071714-58be99b3c4d0
 	github.com/gorilla/mux v1.8.0
 	github.com/hyperledger/aries-framework-go v0.1.7-0.20210816113201-26c0665ef2b9
-	github.com/hyperledger/aries-framework-go-ext/component/storage/couchdb v0.0.0-20210816155437-90525d054756
+	github.com/hyperledger/aries-framework-go-ext/component/storage/couchdb v0.0.0-20210823223913-7a89535c5407
+	github.com/hyperledger/aries-framework-go-ext/component/storage/mongodb v0.0.0-20210823223913-7a89535c5407
 	github.com/hyperledger/aries-framework-go-ext/component/vdr/orb v0.0.0-20210816155124-45ab1ecd4762
-	github.com/hyperledger/aries-framework-go/component/storageutil v0.0.0-20210807121559-b41545a4f1e8
-	github.com/hyperledger/aries-framework-go/spi v0.0.0-20210816113201-26c0665ef2b9
+	github.com/hyperledger/aries-framework-go/component/storageutil v0.0.0-20210823231407-11c14052de4c
+	github.com/hyperledger/aries-framework-go/spi v0.0.0-20210823231407-11c14052de4c
 	github.com/rs/cors v1.7.0
 	github.com/spf13/cobra v1.1.3
 	github.com/stretchr/testify v1.7.0
 	github.com/trustbloc/edge-core v0.1.7-0.20210816120552-ed93662ac716
 	github.com/trustbloc/edv v0.0.0-00010101000000-000000000000
+	golang.org/x/net v0.0.0-20210813160813-60bc85c4be6d // indirect
 )
 
 replace github.com/trustbloc/edv => ../..
