+++ go.mod
@@ -8,16 +8,13 @@ go 1.16
 
 require (
 	github.com/btcsuite/btcutil v1.0.3-0.20201208143702-a53e38424cce
-	github.com/go-kivik/couchdb/v3 v3.2.8 // indirect
 	github.com/google/uuid v1.2.0
 	github.com/gorilla/mux v1.8.0
 	github.com/hyperledger/aries-framework-go v0.1.7-0.20210816113201-26c0665ef2b9
-	github.com/hyperledger/aries-framework-go-ext/component/storage/couchdb v0.0.0-20210816155437-90525d054756
-	github.com/hyperledger/aries-framework-go/component/storageutil v0.0.0-20210807121559-b41545a4f1e8
-	github.com/hyperledger/aries-framework-go/spi v0.0.0-20210816113201-26c0665ef2b9
+	github.com/hyperledger/aries-framework-go/component/storageutil v0.0.0-20210823231407-11c14052de4c
+	github.com/hyperledger/aries-framework-go/spi v0.0.0-20210823231407-11c14052de4c
 	github.com/piprate/json-gold v0.4.1-0.20210813112359-33b90c4ca86c
 	github.com/square/go-jose v2.4.1+incompatible
 	github.com/stretchr/testify v1.7.0
 	github.com/trustbloc/edge-core v0.1.7-0.20210816120552-ed93662ac716
-	golang.org/x/net v0.0.0-20210813160813-60bc85c4be6d // indirect
 )
