--- test/bdd/fixtures/couchdb/.env
@@ -1,16 +1,0 @@
-#
-# Copyright SecureKey Technologies Inc. All Rights Reserved.
-#
-# SPDX-License-Identifier: Apache-2.0
-#
-
-# This file contains the default images and tags used in the docker-based fixtures.
-# The images and tags can be overridden using environment variables.
-# See docker compose documentation.
-
-COUCHDB_IMAGE=couchdb
-COUCHDB_IMAGE_TAG=3.1.0
-COUCHDB_USERNAME=admin
-COUCHDB_PASSWORD=password
-
-COUCHDB_PORT=5984
