--- test/bdd/fixtures/couchdb/docker-compose.yml
@@ -1,25 +1,0 @@
-#
-# Copyright SecureKey Technologies Inc. All Rights Reserved.
-#
-# SPDX-License-Identifier: Apache-2.0
-#
-version: '2'
-
-services:
-
-  couchdb.example.com:
-    container_name: couchdb.example.com
-    image: ${COUCHDB_IMAGE}:${COUCHDB_IMAGE_TAG}
-    ports:
-      - ${COUCHDB_PORT}:${COUCHDB_PORT}
-    environment:
-      - COUCHDB_USER=${COUCHDB_USERNAME}
-      - COUCHDB_PASSWORD=${COUCHDB_PASSWORD}
-    volumes:
-      - ./couchdb-config/config.ini:/opt/couchdb/etc/local.d/config.ini
-    networks:
-      - bdd_net
-
-networks:
-  bdd_net:
-    driver: bridge
