+++ test/bdd/fixtures/edv-rest/.env
@@ -14,6 +14,11 @@ EDV_REST_IMAGE_TAG=latest
 EDV_HOST=0.0.0.0
 EDV_1_PORT=8074
 EDV_2_PORT=8075
-EDV_DATABASE_TYPE=couchdb
-EDV_DATABASE_URL=admin:password@couchdb.example.com:5984
-EDV_DATABASE_PREFIX=edv
+EDV_DATABASE_TYPE=mongodb
+EDV_DATABASE_URL=mongodb://mongodb.example.com:27017
+EDV_DATABASE_PREFIX=edv_
+
+COUCHDB_IMAGE=couchdb
+COUCHDB_IMAGE_TAG=3.1.0
+COUCHDB_USERNAME=admin
+COUCHDB_PASSWORD=password
